+++ previous-versions/sync-for-expenses-version-1/src/sdk/sdk.ts
@@ -55,8 +55,10 @@ export class SDKConfiguration {
     serverDefaults: any;
     language = "typescript";
     openapiDocVersion = "prealpha";
-    sdkVersion = "0.2.0";
-    genVersion = "2.109.3";
+    sdkVersion = "0.3.0";
+    genVersion = "2.144.7";
+    userAgent =
+        "speakeasy-sdk/typescript 0.3.0 2.144.7 prealpha @codat/sync-for-expenses-version-1";
     retryConfig?: utils.RetryConfig;
     public constructor(init?: Partial<SDKConfiguration>) {
         Object.assign(this, init);