+++ previous-versions/sync-for-expenses-version-1/RELEASES.md
@@ -19,3 +19,13 @@ Based on:
 - [typescript v0.2.0] previous-versions/sync-for-expenses-version-1
 ### Releases
 - [NPM v0.2.0] https://www.npmjs.com/package/@codat/sync-for-expenses-version-1/v/0.2.0 - previous-versions/sync-for-expenses-version-1
+
+## 2023-10-04 00:15:50
+### Changes
+Based on:
+- OpenAPI Doc prealpha https://raw.githubusercontent.com/codatio/oas/main/yaml/Codat-Sync-Expenses-v1.yaml
+- Speakeasy CLI 1.93.2 (2.144.7) https://github.com/speakeasy-api/speakeasy
+### Generated
+- [typescript v0.3.0] previous-versions/sync-for-expenses-version-1
+### Releases
+- [NPM v0.3.0] https://www.npmjs.com/package/@codat/sync-for-expenses-version-1/v/0.3.0 - previous-versions/sync-for-expenses-version-1